% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.render_summary_banner}
\alias{.render_summary_banner}
\title{Render the top banner of \code{summary.causal_survival_fit()}}
\usage{
.render_summary_banner()
}
\description{
Render the top banner of \code{summary.causal_survival_fit()}
}
\keyword{internal}
