% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.render_summary_info_box}
\alias{.render_summary_info_box}
\title{Render the method / cohort info box}
\usage{
.render_summary_info_box(est_str, N, K_max, T_max)
}
\description{
Render the method / cohort info box
}
\keyword{internal}
