% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{Values to coalesce.}
}
\value{
\code{x} if not NULL, else \code{y}.
}
\description{
Returns \code{x} if not NULL, otherwise \code{y}.
}
\keyword{internal}
